2009-11-29  Zoltan Varga  <vargaz@gmail.com>
	* debugger-agent.c (resume_thread): Fix a warning.

svn path=/trunk/mono/; revision=147075


Commit migrated from https://github.com/mono/mono/commit/42079727b3aa6527843d7ac0d25f125964c1da88

+++ src/mono/mono/mini/debugger-agent.c
@@ -1890,7 +1890,7 @@ resume_thread (MonoInternalThread *thread)
 
 	g_assert (suspend_count > 0);
 
-	DEBUG(1, fprintf (log_file, "[%p] Resuming thread...\n", (gpointer)thread->tid));
+	DEBUG(1, fprintf (log_file, "[%p] Resuming thread...\n", (gpointer)(gssize)thread->tid));
 
 	tls->resume_count ++;
